# refactor(luasnip): minor code improvements and cleanup