~~- Cache module to avoid repeated require calls~~
- Use more efficient table insertion methods
- Remove dups in cached snippets introduced in commit #712af9f
- Improve code readability