@Saghen 2025-01-23
The `require('luasnip')` calls should be free after the first call: https://www.lua.org/pil/8.1.html. Do you have before and after numbers for the table insertion?